Address review comments for https://github.com/OSGeo/PROJ/pull/1817
