+++ cmake/ProjTest.cmake
@@ -2,6 +2,16 @@
 # add test with sh script
 #
 
+function(proj_test_set_properties TESTNAME)
+    if(MSVC)
+        set_tests_properties( ${TESTNAME}
+            PROPERTIES ENVIRONMENT "PROJ_IGNORE_USER_WRITABLE_DIRECTORY=YES;PROJ_LIB=${PROJECT_BINARY_DIR}/data\\;${PROJECT_SOURCE_DIR}/data")
+    else()
+        set_tests_properties( ${TESTNAME}
+            PROPERTIES ENVIRONMENT "PROJ_IGNORE_USER_WRITABLE_DIRECTORY=YES;PROJ_LIB=${PROJECT_BINARY_DIR}/data:${PROJECT_SOURCE_DIR}/data")
+    endif()
+endfunction()
+
 function(proj_add_test_script_sh SH_NAME BIN_USE)
   if(UNIX)
     get_filename_component(testname ${SH_NAME} NAME_WE)
@@ -26,13 +36,7 @@ function(proj_add_test_script_sh SH_NAME BIN_USE)
         COMMAND ${PROJECT_SOURCE_DIR}/test/cli/${SH_NAME}
         ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${${BIN_USE}}
       )
-      if(MSVC)
-          set_tests_properties( ${testname}
-              PROPERTIES ENVIRONMENT "PROJ_IGNORE_USER_WRITABLE_DIRECTORY=YES;PROJ_LIB=${PROJECT_BINARY_DIR}/data\\;${PROJECT_SOURCE_DIR}/data")
-      else()
-          set_tests_properties( ${testname}
-              PROPERTIES ENVIRONMENT "PROJ_IGNORE_USER_WRITABLE_DIRECTORY=YES;PROJ_LIB=${PROJECT_BINARY_DIR}/data:${PROJECT_SOURCE_DIR}/data")
-      endif()
+      proj_test_set_properties(${testname})
     endif()
 
   endif()
@@ -49,13 +53,6 @@ function(proj_add_gie_test TESTNAME TESTCASE)
       ${TESTFILE}
     )
 
-    if(MSVC)
-        set_tests_properties( ${TESTNAME}
-            PROPERTIES ENVIRONMENT "PROJ_IGNORE_USER_WRITABLE_DIRECTORY=YES;PROJ_LIB=${PROJECT_BINARY_DIR}/data\\;${PROJECT_SOURCE_DIR}/data")
-    else()
-        set_tests_properties( ${TESTNAME}
-            PROPERTIES ENVIRONMENT "PROJ_IGNORE_USER_WRITABLE_DIRECTORY=YES;PROJ_LIB=${PROJECT_BINARY_DIR}/data:${PROJECT_SOURCE_DIR}/data")
-    endif()
-
+    proj_test_set_properties(${TESTNAME})
 
 endfunction()
