+++ data/proj.ini
@@ -11,6 +11,6 @@ cdn_endpoint = https://cdn.proj.org
 
 cache_enabled = on
 
-cache_size_MB = 100
+cache_size_MB = 300
 
 cache_ttl_sec = 86400
