+++ src/filemanager.hpp
@@ -88,6 +88,10 @@ class File {
     const std::string &name() const { return name_; }
 };
 
+// ---------------------------------------------------------------------------
+
+std::unique_ptr<File> pj_network_file_open(PJ_CONTEXT* ctx, const char* filename);
+
 NS_PROJ_END
 
 //! @endcond Doxygen_Suppress
